Sound editions by PH-COR
Submitted via email:
I have reviewed the  sound of the A380-omega. Attached a Sounds directory in Zip format to replace the present one. The modifications are more than minor and checkflights completed the balancing of the soundlevels.

the typical A380 rotor-whine is better simulated. Touch down squeals are heard outside and touchdown bonks are heard inside the cockpit. Runway rumble and bump are simulated wind noises and nosegear locking sounds that can be heard in the cockpit above the nosegearbay are simulated.  V1 callout at startup is suppressed